requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.1
django-environ==0.10.0
djangorestframework==3.14.0
idna==3.4
mpmath==1.3.0
oauthlib==3.2.2
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
sympy==1.12
-e git+https://github.com/yellow-carrot/ToDo_List.git@0ac4a02986e3fad45d63ea552ccbd6ebcd33bc49#egg=todolist
urllib3==2.0.3
